@@ -1,16 +1,19 @@
 {
     "name": "fabric-wallet-migration",
-    "version": "0.1.0",
+    "version": "0.2.0",
     "description": "Migration from Hyperledger Fabric 1.4 to 2.0 wallets",
     "main": "index.js",
     "types": "index.d.ts",
+    "engines": {
+        "node": ">=10.13.0"
+    },
     "scripts": {
-        "build": "npm run compile && npm run lint",
-        "clean": "rm -rf dist *.tgz scenario/package-lock.json",
+        "build": "npm-run-all clean compile lint",
+        "clean": "rimraf dist *.tgz scenario/package-lock.json scenario/node_modules",
         "compile": "tsc",
         "lint": "eslint . --ext .ts",
         "scenario": "./runScenario.sh",
-        "test": "npm run unitTest && npm run scenario",
+        "test": "npm-run-all unitTest scenario",
         "unitTest": "jest"
     },
     "repository": {
@@ -27,19 +30,20 @@
     },
     "homepage": "https://github.com/bestbeforetoday/fabric-wallet-migration#readme",
     "devDependencies": {
-        "@types/jest": "^25.1.3",
-        "@types/rimraf": "^2.0.3",
-        "@types/uuid": "^7.0.0",
-        "@typescript-eslint/eslint-plugin": "^2.22.0",
-        "@typescript-eslint/parser": "^2.22.0",
-        "eslint": "^6.8.0",
-        "fabric-network": "^1.4.7",
-        "jest": "^25.1.0",
-        "ts-jest": "^25.2.1",
-        "typescript": "^3.8.3"
+        "@types/jest": "^26.0.9",
+        "@types/rimraf": "^3.0.0",
+        "@types/uuid": "^8.3.0",
+        "@typescript-eslint/eslint-plugin": "^3.9.0",
+        "@typescript-eslint/parser": "^3.9.0",
+        "eslint": "^7.6.0",
+        "fabric-network": "^1.4.11",
+        "jest": "^26.4.0",
+        "npm-run-all": "^4.1.5",
+        "ts-jest": "^26.2.0",
+        "typescript": "^3.9.7"
     },
     "dependencies": {
         "rimraf": "^3.0.2",
-        "uuid": "^7.0.1"
+        "uuid": "^8.3.0"
     }
 }
