@@ -2,11 +2,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import fs = require("fs");
-import util = require("util");
-import path = require("path");
-import os = require("os");
-import _rimraf = require("rimraf");
+import * as fs from "fs";
+import * as util from "util";
+import * as path from "path";
+import * as os from "os";
+import * as _rimraf from "rimraf";
 const rimraf = util.promisify(_rimraf);
 
 export async function readFile(fileName: string): Promise<string> {
